# api/openapi.yaml
openapi: 3.1.0
info:
  title: BasePod API
  version: "0.5.0"
  description: |
    BasePod's control-plane REST API — session auth, app lifecycle, deploys
    (image/tarball/git/compose), env/domains, and streaming logs/build-log/
    stats over Server-Sent Events.

    **This document is hand-written and enforced, not generated.** A test
    (`internal/api/openapi_test.go`) walks the real chi router with
    `chi.Walk` and asserts every registered route+method appears here, and
    every path+method here exists in the router — in both directions. If
    this file and the router ever disagree, that test fails. Treat any
    change to `internal/api/router.go`'s route table as also owing a change
    here.

    **Error envelope.** Every non-2xx JSON response uses the same shape:
    `{"error":{"code":"...","message":"..."}}` — see the shared `Error`
    schema referenced from every operation's error responses below.

    **Auth.** Two independent schemes exist:
    - `sessionAuth` — `Authorization: Bearer <session-token>`, minted by
      `POST /auth/login` and valid up to 30 days. Used by every
      authenticated route except the two noted below.
    - `streamTokenAuth` — `?access_token=<stream-token>` on the query
      string, accepted **only** on the three SSE routes
      (`.../logs`, `.../deployments/{number}/log`, `.../stats`), because a
      native browser `EventSource` cannot set request headers. A stream
...
      (it is single-use-scope and short-lived).
...
servers:
...
security:
  - sessionAuth: []
...
paths:
...
  /stream-token:
    post:
      operationId: createStreamToken
      tags: [streaming]
      summary: Mint a short-lived, single-purpose SSE stream token
      description: |
        Scoped to exactly one (scope, slug[, deployment_number]) triple —
        see the top-level auth description for how it's used on the three
        SSE routes. Expires in 5 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StreamTokenRequest" }
      responses:
        "200":
          description: Minted token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamTokenResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Validation" }
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      description: "Session token from POST /auth/login, sent as `Authorization: Bearer <token>`."
    streamTokenAuth:
      type: apiKey
      in: query
      name: access_token
      description: |
        A stream token from POST /stream-token, scoped to one SSE route's
        (scope, slug[, deployment_number]) — see the top-level auth
        description. Accepted ONLY on the three SSE routes; a session
        token here is rejected everywhere, including on those routes.
...
  responses:
    Unauthorized:
      description: Missing, malformed, or expired credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Validation:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: "Machine-readable error code, e.g. \"app_not_found\"." }
            message: { type: string }
...
    StreamTokenRequest:
      type: object
      required: [scope, slug]
      properties:
        scope: { type: string, enum: [app_logs, build_log, stats] }
        slug: { type: string }
        deployment_number:
          type: integer
          description: Required for scope "build_log", must be absent otherwise.
    StreamTokenResponse:
      type: object
      properties:
        token: { type: string }
        expires_at: { type: string, format: date-time }
...